Big blob commit
This is far from ideal, but between my messing up github and my
abandoning substantial work and forgetting the details, it's hard to
generate anything clear.

We've created some abstractions with unit tests and generated to models.
One is a end-to-end test of project creation behavior to demonstrate the
statistics of the model project.s The other tests the how coding
practices with resulting in long files on average affect the amount of
manual labor required for code merges.